Updated the testing framework to have a FIXVersionedTestSuite that runs the specified tests through all the specified FIX versions. Also takes in a list of exclusions, to run those tests only under different set of versions. Works with the FIXVersionTestCase abstract superclass.
Also introduced more functionality to the FIXMessageAugmentor framework, moving version-specific message setting code there from FIXMessageFactory.

As a result, the unit tests now have to be special-cased in a few places, to make sure we don't check for fields in specific FIX versions where they are not defined yet. For example, LeavesQty is only applicable in FIX41+



git-svn-id: file:///space/source.marketcetera.org/root/trunk@4452 501a4a96-914b-0410-8454-f5988a4ca596