Fix for bug #146: right-justify numbers in tradebase
Introduced a new number_format_helper that has a function nf() that takes in a number (string or a BigDecimal), rounds it ot the specified # of decimal places, drops the decimals if needed and displays it (padding with extra decimal 0s if needed).  It can be simplified greatly in the future once we can start using Ruby 1.8.5 and can use BigDecimal.limit() function.

moved a few other display-related helper functions to number_format_helper and got rid of display_helper



git-svn-id: file:///space/source.marketcetera.org/root/trunk@8211 501a4a96-914b-0410-8454-f5988a4ca596