Changes related to getting OptionOrderTicketViewTest to work. See related bug #307
Modified IMarketDataListCallback to extend IMarketDataListener, removing the onMarketDataListAvailable() function. We should be using the IMarketDataListener onMessages() function instead.

Modified other files to reflect this change.

Decomped OptionOrderTicketControllerHelper to have a handleMarketDataList() function that can be called from the outside (tests) to fake incoming market quotes

OptionOrderTicketControllerHelper: disabled the listener that used to listen on changes to getOptionSymbolControl. that seems to fix the stacktrace overrun in testShowQuote but wanted to double-check with you about this. (line 121)

Made AbstractOrderTicketController implement IMarketDataListener, updated related classes



git-svn-id: file:///space/source.marketcetera.org@6733 501a4a96-914b-0410-8454-f5988a4ca596