Added isTradingSessionStatus() helper to FIXMessageUtil and test Added logging call in QuickFIXApplication to print when incoming trading session comes in.
Created an OMSAdmin MBean that allows you to send a UserRequest message that resets a password
this can be enacted from jconsole or any other JMX management interface, for example.

Made the IDFactory an object that's created through Spring, and set that in OutgoingMessageHandler and OMSAdmin via Spring.
As a result, had to move the jdbc config params into their own beans in oms.xml file


git-svn-id: file:///space/source.marketcetera.org/root/trunk@3373 501a4a96-914b-0410-8454-f5988a4ca596