Fix for EG-320, adjusts for long-running start and stop methods in strategies.  Specific changes as follows:
This change addresses EG-320. The problem stated therein was that if a strategy implemented either "onStart" or "onStop" that took an arbitrarily long time to complete, the module 
framework would be locked for the duration. For the pathological case, where the strategy method never completes, the module framework would remain locked indefinitely.

Eventually, the solution will probably be to modify the module framework such that it does not lock during module start and stop. For now, though, and possibly going forward as well 
depending on the final resolution of the module framework, the strategy onStart and onStop will be executed asynchronously.

The specific changes covered by this review are as follows:

1) Add getStatus() to the strategy's MXBean interface. This returns the status of the underlying strategy object, not the status of the strategy module which is available from the 
module framework.
2) Modify AbstractExecutor to execute onStart and onStop asynchronously
3) Add more consistency and complexity to the strategy status attribute. There are now more states and the transitions between them are defined and restricted. I will be adding more 
documentation to cover the state changes.
4) Add restrictions to outgoing and incoming data requests and objects based on the state of the strategy. A strategy that is not started, failed, stopping, or stopped may not 
receive or request data
5) Modify the callback service to make sure that a strategy that is stopping or stopped may not request callbacks
6) Add interrupt() to MXBean interface - this allows a strategy to be interrupted if it does not respond to requests to stop.  The interrupt operation will try its best to halt a 
running strategy.  It may safely be called on a strategy in any state, but will always try to move the strategy to STOPPED.
7) Add JMX NotificationEmitter service to strategy module.  The JMX notifier will publish changes to the strategy status to subscribers.  The StrategyModule has a static accessor 
that will return a JMXInterface for a given strategy URN.
8) Unit tests that test the above (note that the interrupt test was added for java strategies only as I haven't yet figured out the JRuby thread interruption semantics)

All unit tests pass.

Reviewed as ENG-198


git-svn-id: file:///space/source.marketcetera.org/root/trunk@10201 501a4a96-914b-0410-8454-f5988a4ca596