Author: tlerios
Reviewer: will

Crucible: ENG-197

JIRA: EG-327 (part 3)

Comments:

This is part of EG-327: the ORS contains several classes and config
items that can be transitioned before the rest of the code is. This
gradual transition can reduce the workload on Monday, and minimize
transition risk (easier to identify the cause of a build failire).

There are more changes I will make before Monday, but those will go in
a branch.




git-svn-id: file:///space/source.marketcetera.org/root/trunk@10203 501a4a96-914b-0410-8454-f5988a4ca596