NYSE SuperFeed Marketdata Adapter (reviewed by Will as ENG-447)
This review covers the initial iteration of the NYSE SuperFeed Marketdata Adapter.  This review is the public side; the changes here are minimal.  I added the 
UriScheme class by refactoring this code out of the OpenTick adapter.  I did not retrofit the OpenTick adapter because it's very shortly going to be removed from the 
codebase.  The utility of UriScheme is that it provides an easy way to construct custom URIs to encapsulate credentials ("nyse://hostname:port").  This is a common 
pattern that shows up in adapters, so there's reason to make it general.

I ran unit tests on Linux and Windows plus lots of manual testing.


git-svn-id: file:///space/source.marketcetera.org@10632 501a4a96-914b-0410-8454-f5988a4ca596