Fix for STG-9 Strategy cancelReplace order returns the wrong OrderID (reviewed as ENG-449 by Will)
This was reported on the forum (http://www.marketcetera.org/community/posts/list/0/209.page#650).  When creating a cancel/replace, the strategy API returns the 
orderID of the replacing order rather than the orderID of the replace request itself.  Since the orderID of the replace request is the only piece of information that 
caller doesn't have, it makes more sense to return that.

Manual reproduction of the original scenario passes as do all strategy unit tests.



git-svn-id: file:///space/source.marketcetera.org@10633 501a4a96-914b-0410-8454-f5988a4ca596