NYSE Adapter Refactoring (ENG-451 reviewed by Will)
This change is a small NYSE adapter refactoring. It seeks to accomplish the following objectives:

1) It turns out that the NYSE sequence number is not unique because a single NYSE update can spawn more than one EventBase object. Since the event contract guarantees 
that every event must have a unique message ID (within the same JRE instance), the sequence number is ill-suited for message ID. Therefore, a unique message ID is 
assigned to the events. Since the sequence number is required to be visible on the events, new NYSE-specific event subclasses were created that make the sequence 
number available.

2) Cosmetic refactoring of NyseEventTranslator to reduce code redundancy. Some of the processing was shifted to the NyseMarketDataFeed itself at the point of message 
receipt.

3) Cosmetic refactoring of NyseEventTranslator to ensure the event translator is always a singleton object. This was essentially a stylistic change.


git-svn-id: file:///space/source.marketcetera.org@10635 501a4a96-914b-0410-8454-f5988a4ca596