This change covers several bugs in the Strategy API all related to how the Strategy API caches submitted orders and tracks execution reports. The purpose of order and ER caching is to allow the Stratey API to track the orders (OrderSingle, OrderReplace, and OrderCancel) that are submitted during a strategy session and the execution reports (ERs) that are received indicating the status of the orders.
The following bugs are addressed: EG-667, EG-762, EG-793, EG-794, and EG-798.  The new behavior is as follows:

1) Unrelated to the caching issues, Strategy API operations that *send* data (as opposed to ones that receive data) are now allowed in onStop 
(EG-667)
2) Orders and OrderIDs for orders submitted during a strategy session are now tracked only until the order is no longer open.  Previously, they 
were tracked only until a cancel was issued.  Since a cancel may be rejected and need to be reissued, the previous logic was faulty.  While 
more complicated, the current logic will continue to allow orders to be canceled/replaced until the order is proven to be closed by an ER 
(EG-793, EG-794)
3) The BrokerOrderID is intentionally unspecified on outgoing cancel and replace orders (EG-762)
4) The price is intentionally unspecified on outgoing replace orders where the original order was of type Market (EG-798)

Strategy unit tests pass as does manual testing of each reproduction scenario.


git-svn-id: file:///space/source.marketcetera.org/root/trunk@10767 501a4a96-914b-0410-8454-f5988a4ca596